+++ service/grails-app/domain/org/olf/erm/HoldingsCoverage.groovy
@@ -1,47 +1,48 @@
 package org.olf.erm
 
+import java.time.LocalDate
+
+import org.olf.kb.AbstractCoverageStatement
+
 import grails.gorm.MultiTenant
 
 
 /**
- * A coverage statement - 
+ * Represents the local coverage override
  */
-public class HoldingsCoverage implements MultiTenant<HoldingsCoverage> {
-
+public class HoldingsCoverage extends AbstractCoverageStatement implements MultiTenant<HoldingsCoverage> {
+  
   String id
 
-  // Mutually exclusive --- ONE of pci, pti or ti
-  Entitlement entitlement
-
-  // MUST Be in format yyyy-mm-dd
-  String startDate
-  String endDate
+  LocalDate startDate
+  LocalDate endDate
+  
   String startVolume
   String startIssue
   String endVolume
   String endIssue
-
+  
+  static belongsTo = [entitlement: Entitlement]
+  
   static mapping = {
-                   id column:'co_id', generator: 'uuid', length:36
-              version column:'co_version'
-          entitlement column:'co_ent_fk'
-            startDate column:'co_start_date'
-              endDate column:'co_end_date'
-          startVolume column:'co_start_volume'
-           startIssue column:'co_start_issue'
-            endVolume column:'co_end_volume'
-             endIssue column:'co_end_issue'
+             id column:'co_id', generator: 'uuid', length:36
+        version column:'co_version'
+    entitlement column:'co_ent_fk'
+      startDate column:'co_start_date'
+        endDate column:'co_end_date'
+    startVolume column:'co_start_volume'
+     startIssue column:'co_start_issue'
+      endVolume column:'co_end_volume'
+       endIssue column:'co_end_issue'
   }
-
+  
   static constraints = {
-    entitlement(nullable:true, blank:false);
-    startDate(nullable:true, blank:false);
-    endDate(nullable:true, blank:false);
-    startVolume(nullable:true, blank:false);
-    startIssue(nullable:true, blank:false);
-    endVolume(nullable:true, blank:false);
-    endIssue(nullable:true, blank:false);
+    entitlement(nullable:false)
+    startDate(nullable:false, validator: STATEMENT_START_VALIDATOR)
+    endDate(nullable:true)
+    startVolume(nullable:true, blank:false)
+    startIssue(nullable:true, blank:false)
+    endVolume(nullable:true, blank:false)
+    endIssue(nullable:true, blank:false)
   }
-
-
 }