+++ service/grails-app/domain/org/olf/kb/ErmResource.groovy
@@ -22,10 +22,12 @@ public class ErmResource implements MultiTenant<ErmResource> {
   Date lastUpdated
   
   static hasMany = [
+    coverage: CoverageStatement,
     entitlements: Entitlement
   ]
 
   static mappedBy = [
+    coverage: 'resource',
     entitlements: 'resource'
   ]
   static mapping = {
@@ -37,6 +39,7 @@ public class ErmResource implements MultiTenant<ErmResource> {
           dateCreated column:'res_date_created'
           lastUpdated column:'res_last_updated'
               subType column:'res_sub_type_fk'
+             coverage cascade: 'all-delete-orphan'
   }
 
   static constraints = {
@@ -46,6 +49,7 @@ public class ErmResource implements MultiTenant<ErmResource> {
          subType (nullable:true, blank:false)
      dateCreated (nullable:true, blank:false)
      lastUpdated (nullable:true, blank:false)
+      coverage (validator: CoverageStatement.STATEMENT_COLLECTION_VALIDATOR, sort:'startDate')
   }
   
   String toString() {